// src/schemas/public/Inventory.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ProductsId } from './Products';

/** Identifier type for public.inventory */
export type InventoryId = number & { __brand: 'InventoryId' };

/** Represents the table public.inventory */
export default interface Inventory {
  id: InventoryId;

  product_id: ProductsId;

  quantity: number;
}

/** Represents the initializer for the table public.inventory */
export interface InventoryInitializer {
  /** Default value: nextval('inventory_id_seq'::regclass) */
  id?: InventoryId;

  product_id: ProductsId;

  quantity: number;
}

/** Represents the mutator for the table public.inventory */
export interface InventoryMutator {
  id?: InventoryId;

  product_id?: ProductsId;

  quantity?: number;
}