// src/schemas/public/ProductRatings.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the materializedView public.product_ratings */
export default interface ProductRatings {
  product_id: number | null;

  one_star: number | null;

  two_stars: number | null;

  three_stars: number | null;

  four_stars: number | null;

  five_stars: number | null;

  total_reviews: number | null;

  average_rating: number | null;
}